Move shared FromAddress methods into a FromAddressObjects module
These methods are now used in two controllers. Move the common methods
out into a shared module to dry up the code.
